Disable outdated crypto, provide DH params for PFS
To be designated "modern" by Chrome, it needs to be able to negotiate:

  * (EC)DHE key exchange (for PFS)
  * AES or CHACHA cipher
  * AEAD MAC

Everything else is considered "obsolete".

Source: IsSecureTLSCipherSuite in
https://chromium.googlesource.com/chromium/src/+/master/net/ssl/ssl_cipher_suite_names.cc

DH key exchange requires parameters (a long prime). These can be generated
with "openssl dhparam 2048" and should ideally be provided by the user,
so we try to get them from the cert+key bundle file first, but fall back
on hard-coded ones if none are supplied.

Fixes #4